pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.10",
    "google-genai>=1.16.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5",
]
